+++ python/fate/components/components/__init__.py
@@ -13,27 +13,26 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-# from .dataframe_io_test import dataframe_io_test
-# from .dataframe_transformer import dataframe_transformer
+from .dataframe_io_test import dataframe_io_test
+from .dataframe_transformer import dataframe_transformer
 from .demo import run
-
-# from .evaluation import evaluation
-# from .feature_scale import feature_scale
-# from .hetero_lr import hetero_lr
-# from .intersection import intersection
-# from .multi_model_test import multi_model_test
-# from .reader import reader
-# from .statistics import statistics
+from .evaluation import evaluation
+from .feature_scale import feature_scale
+from .hetero_lr import hetero_lr
+from .intersection import intersection
+from .multi_model_test import multi_model_test
+from .reader import reader
+from .statistics import statistics
 
 BUILDIN_COMPONENTS = [
-    # hetero_lr,
-    # reader,
-    # feature_scale,
-    # intersection,
-    # evaluation,
+    hetero_lr,
+    reader,
+    feature_scale,
+    intersection,
+    evaluation,
     run,
-    # dataframe_transformer,
-    # statistics,
-    # dataframe_io_test,
-    # multi_model_test,
+    dataframe_transformer,
+    statistics,
+    dataframe_io_test,
+    multi_model_test,
 ]
