+++ python/fate/interface/__init__.py
@@ -15,21 +15,14 @@
 from ._address import Address
 from ._cipher import CipherKit, PHECipher
 from ._computing import ComputingEngine
-from ._consts import T_ARBITER, T_GUEST, T_HOST, T_ROLE
 from ._data_io import Dataframe
 from ._federation import FederationEngine, FederationWrapper
 from ._gc import GarbageCollector
-from ._metric import InCompleteMetrics, Metric, Metrics, MetricsHandler, MetricsWrap
 from ._party import Parties, Party, PartyMeta
 from ._table import CSessionABC, CTableABC
 
 __all__ = [
     "Dataframe",
-    "MetricsHandler",
-    "MetricsWrap",
-    "Metrics",
-    "InCompleteMetrics",
-    "Metric",
     "Party",
     "Parties",
     "PartyMeta",
@@ -39,8 +32,4 @@ __all__ = [
     "PHECipher",
     "FederationEngine",
     "GarbageCollector",
-    "T_GUEST",
-    "T_HOST",
-    "T_ARBITER",
-    "T_ROLE",
 ]
