--- python/fate/interface/_consts.py
@@ -1,20 +1,0 @@
-#
-#  Copyright 2019 The FATE Authors. All Rights Reserved.
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
-from typing import Literal
-
-T_GUEST = Literal["guest"]
-T_HOST = Literal["host"]
-T_ARBITER = Literal["arbiter"]
-T_ROLE = Literal[T_GUEST, T_HOST, T_ARBITER]
